+++ src/device/common.h
@@ -408,11 +408,6 @@ struct RunWorkColl {
 template<ncclFunc_t Fn, typename T, typename RedOp, int Algo, int Proto, int COLL_UNROLL>
 struct RunWorkBatch;
 
-#ifndef RCCL_EXPOSE_STATIC
-// Specialized for P2p in sendrecv.h
-template<typename T, typename RedOp>
-struct RunWorkBatch<ncclFuncSendRecv, T, RedOp, NCCL_ALGO_RING, NCCL_PROTO_SIMPLE>;
-#endif
 // Specialized here for non-P2p (Coll and CollReg)
 template<ncclFunc_t Fn, typename T, typename RedOp, int Algo, int Proto, int COLL_UNROLL>
 struct RunWorkBatch {
