--- src/misc/rccl_wrap.cc
@@ -1,39 +1,0 @@
-/* ************************************************************************
- * Copyright (C) 2016-2025 Advanced Micro Devices, Inc. All rights reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell cop-
- * ies of the Software, and to permit persons to whom the Software is furnished
- * to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IM-
- * PLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNE-
- * CTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- * ************************************************************************ */
-
-#include "rccl_wrap.h"
-
-extern size_t ncclFuncMaxSendRecvCount(ncclFunc_t func, int nRanks, size_t count);
-extern ncclResult_t getAlgoInfo(
-    struct ncclComm* comm, struct ncclTaskColl* task,
-    int collNetSupport, int nvlsSupport, int numPipeOps, ncclSimInfo_t* simInfo = NULL
-);
-
-ncclResult_t rcclGetAlgoInfo(
-    struct ncclComm* comm, struct ncclTaskColl* task,
-    int collNetSupport, int nvlsSupport, int numPipeOps, ncclSimInfo_t* simInfo) {
-    return getAlgoInfo(comm, task, collNetSupport, nvlsSupport, numPipeOps, simInfo);
-}
-
-size_t rcclFuncMaxSendRecvCount(ncclFunc_t func, int nRanks, size_t count) {
-    return ncclFuncMaxSendRecvCount(func, nRanks, count);
-}
